affichage des livrables

+++ src/app/projets/livrable/livrable.component.ts
@@ -1,5 +1,8 @@
 import { Component } from '@angular/core';
+import { FormBuilder, FormGroup } from '@angular/forms';
 import { Router } from '@angular/router';
+import { Observable, catchError, throwError } from 'rxjs';
+import { Livrable } from 'src/app/model/livrable';
 import { LivrableService } from 'src/app/services/livrable.service';
 
 @Component({
@@ -8,25 +11,37 @@ import { LivrableService } from 'src/app/services/livrable.service';
   styleUrls: ['./livrable.component.css'],
 })
 export class LivrableComponent {
-  livrables!: any[];
+  livrables!: Observable<Array<Livrable>>;
+  errorMessage!: string;
+  searchFormGroup: FormGroup | undefined;
   constructor(
     private livrableService: LivrableService,
-    private router: Router
+    private router: Router,
+    private fb: FormBuilder
   ) {}
   ngOnInit(): void {
-    this.getLivrables();
-  }
-  getLivrables() {
-    this.livrableService.getLivrables().subscribe((data) => {
-      this.livrables = data;
+    this.searchFormGroup = this.fb.group({
+      keyword: this.fb.control(''),
     });
+    this.handleSearchLivrables();
   }
   supprimerLivrable(id: number) {
+    let conf = confirm('Êtes-vous sûr de vouloir supprimer ce livrable ?');
+    if (!conf) return;
     this.livrableService.deleteLivrable(id).subscribe(() => {
-      this.getLivrables();
+      this.handleSearchLivrables();
     });
   }
   ajouterLivrable() {
     this.router.navigateByUrl('/admin/addlivrable');
+  }
+  handleSearchLivrables() {
+    let kw = this.searchFormGroup?.value.keyword;
+    this.livrables = this.livrableService.searchLivrable(kw).pipe(
+      catchError((err) => {
+        this.errorMessage = err.message;
+        return throwError(err);
+      })
+    );
   }
 }
